Lukej/aircop fix2 (#86)
* add cudnn relu, elu, tanh and sigmoid to unary frontend

* move softmax to misc

* add misc

* move cudnn activation to unary

* clean up
