+++ src/main/scala/lms/transformation/distributed_tensor/distributedTensorUnary.scala
@@ -11,7 +11,8 @@ import lms.thirdparty.array_computation.{ArrayCPUOps, CUDATypeLess, CudaOps}
 
 import Backend._
 
-trait FixedSizeDistributedTensorUnaryTypeLess extends FixedSizeDistributedTensorMutationTypeLess with FixedSizeDistributedTensorBinaryTypeLess {
+trait FixedSizeDistributedTensorUnaryTypeLess extends FixedSizeDistributedTensorMutationTypeLess 
+  with FixedSizeDistributedTensorBinaryTypeLess with FixedSizeDistributedTensorConvTypeLess {
 
   def Transpose(tensor: TENSOR, anno: Anno = NAnno)(implicit __pos: SourceContext): TENSOR = {
     assert(tensor.shapeSize.size == 2, "input of transpose must be 2D")
@@ -39,10 +40,23 @@ trait FixedSizeDistributedTensorUnaryTypeLess extends FixedSizeDistributedTensor
     (new TENSOR(Adapter.g.reflectRead("tensor_relu", C(res_tt), C(anno), tensor.x)(tensor.x)).withSrcType(__pos, tensor.et))
   }
 
+  def ActivationForward(input: TENSOR, params: ActivationParam, mode: String, anno: Anno, __pos: SourceContext): TENSOR = {
+    val res_tt = input.resultType
+    (new TENSOR(Adapter.g.reflectRead("tensor_activation", C(res_tt), C(anno), input.x, 
+      C(params), C(mode))(input.x)).withSrcType(__pos, input.et))
+  }
+
+  def ActivationBackward(input: TENSOR, output: TENSOR, doutput: TENSOR, params: ActivationParam, mode: String, anno: Anno, __pos: SourceContext): TENSOR = {
+    val res_tt = doutput.resultType
+    (new TENSOR(Adapter.g.reflectRead("tensor_activation_bwd", C(res_tt), C(anno), input.x, output.x, doutput.x, 
+      C(params), C(mode))(input.x, output.x, doutput.x)).withSrcType(__pos, doutput.et))
+  }
+
   val unaryops = List("tensor_transpose", "tensor_negate", "tensor_invert", "tensor_tanh", "tensor_relu")
 
   override def mergable_dims(node: Node) = node match {
     case Node(s, op, _, _) if (unaryops.contains(op))=> List()
+    case Node(s, "tensor_activation", _, _) => List()
     case _ => super.mergable_dims(node)
   }
 
@@ -94,6 +108,16 @@ trait FixedSizeDistributedTensorUnaryTypeLess extends FixedSizeDistributedTensor
         (() => {
           Accumulate(gradMap(a), ReluGrad(gradMap(s), new TENSOR(transform(a)), anno), anno); ()
         }) +=: backwardNodes
+    
+    case Node(s, "tensor_activation", tt::Backend.Const(anno:Anno)::(a:Backend.Sym)::Backend.Const(params:ActivationParam)::Backend.Const(mode:String)::_, _) =>
+        implicit val pos = Adapter.oldSourceMap(s)
+        forwardNodes += node
+        (() => {
+            val x = new TENSOR(transform(a))
+            val y = new TENSOR(transform(s))
+            val grad = ActivationBackward(x, y, gradMap(s), params, mode, anno, pos)
+            Accumulate(gradMap(a), grad, anno); ()
+        }) +=: backwardNodes
 
     case _ => super.aircopCollect(node, forwardNodes, weightNodes, backwardNodes, gradMap, momentumMap, transform)
   }
@@ -128,6 +152,35 @@ trait FixedSizeDistributedTensorOpsUnary extends FixedSizeDistributedTensorOpsBa
 
     def relu(anno: Anno)(implicit __pos: SourceContext): Rep[Tensor[T]] = {
       val t = Relu(self, anno)
+      Wrap[Tensor[T]](t.x)
+    }
+
+    // clipped relu
+    def relu(threshold: Float, anno: Anno)(implicit __pos: SourceContext): Rep[Tensor[T]] = {
+      val self = tensor(x)
+      val p = ActivationParam(1.0f, 0.0f, threshold)
+      val t = ActivationForward(self, p, "crelu", anno, __pos)
+      Wrap[Tensor[T]](t.x)
+    }
+
+    def elu(alpha: Float, anno: Anno)(implicit __pos: SourceContext): Rep[Tensor[T]] = {
+      val self = tensor(x)
+      val p = ActivationParam(1.0f, 0.0f, alpha)
+      val t = ActivationForward(self, p, "elu", anno, __pos)
+      Wrap[Tensor[T]](t.x)
+    }
+
+    def sigmoid(anno: Anno)(implicit __pos: SourceContext): Rep[Tensor[T]] = {
+      val self = tensor(x)
+      val p = ActivationParam(1.0f, 0.0f, 0.0f)
+      val t = ActivationForward(self, p, "sigmoid", anno, __pos)
+      Wrap[Tensor[T]](t.x)
+    }
+    
+    def cudnn_tanh(anno: Anno)(implicit __pos: SourceContext): Rep[Tensor[T]] = {
+      val self = tensor(x)
+      val p = ActivationParam(1.0f, 0.0f, 0.0f)
+      val t = ActivationForward(self, p, "tanh", anno, __pos)
       Wrap[Tensor[T]](t.x)
     }
   }
